+++ modules/hanna-react/src/utils.ts
@@ -1,6 +1,6 @@
 export * from './utils/browserSide.js';
 export * from './utils/config.js';
-export * from './utils/HannaUIState.js';
+export { HannaUIState, useHannaUIState } from './utils/HannaUIState.js';
 export * from './utils/useDidChange.js';
 export * from './utils/useFormatMonitor.js';
 export * from './utils/useGetSVGtext.js';